Moves based on merge
