+++ site/src/pages/login.tsx
@@ -1,10 +1,10 @@
 import { makeStyles } from "@material-ui/core/styles"
 import { useActor } from "@xstate/react"
 import React from "react"
-import { userXService } from "../xServices/user/userXService"
 import { SignInForm } from "./../components/SignIn"
 import { Navigate, useLocation } from "react-router-dom"
 import { Location } from "history"
+import { userXService } from "../../xServices/user/userXService"
 
 export const useStyles = makeStyles((theme) => ({
   root: {
